refactor: extract message/media UI into MessageMediaViews.swift
Move the message rendering and media subsystem (ConversationMessageRow,
TimelineNoticeRow, MessageBubble, visual-media grids/tiles, attachment
rows, audio/video players, image-gallery overlay, and inline message
actions/reactions) out of MessengerShellView.swift into
MessageMediaViews.swift. Pure code movement: top-level declarations
promoted to internal; the Composer audio-waveform types
(ComposerAudioWaveformView/Mode/Bar/Presentation, still in the shell file)
promoted to internal so the message audio player can reference them. No
logic or behavior changes.

Build green; test suite unchanged (same two pre-existing image-downsampling
failures, unrelated to view extraction).
